+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix Vanilla Additions Config not working (Fabric Only)
 - Add Blueberries to Fox Food
 - Allow placing BWG Saplings on all FarmLand Blocks
+- Allow planting BushBlock on all FarmLand Blocks (Fabric Only)
 
 # 1.6.4
 - Add Wood Blewit to Enderman Holdable BlockTag